amountforalllast

+++ src/main/java/com/crypto/service/WalleteServiceImplementation.java
@@ -31,7 +31,7 @@ public class WalleteServiceImplementation implements WalletService {
     public Wallet genrateWallete(User user) {
         Wallet wallet=new Wallet();
         wallet.setUser(user);
-        wallet.setBalance(10000);
+        wallet.setBalance(BigDecimal.valueOf(10000));
         return walletRepository.save(wallet);
     }
